Linux network scanner improvements
After testing the software on Ubuntu, a network scan was being run on an incorrect and assumed network device. Now the software will find the real network devices and run scans on them. Also, if the application is not run with super user rights, it will try to ask for them using an askpass program if provided in SUDO_ASKPASS environment variable.